+++ src/include/LSP/Sourcemap.hpp
@@ -19,6 +19,7 @@ struct SourceNode
     bool isScript();
     std::optional<std::filesystem::path> getScriptFilePath();
     Luau::SourceCode::Type sourceCodeType() const;
+    std::optional<SourceNodePtr> findChild(const std::string& name);
 };
 
 static void from_json(const json& j, SourceNode& p)